replaced previous HKT processor
