+++ kind/processor/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,2 +1,2 @@
-com.github.tonivade.purefun.NewHigherKindProcessor
+com.github.tonivade.purefun.HigherKindProcessor
 com.github.tonivade.purefun.InstanceProcessor
